LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-percent-size-child-expected.txt

Test that percentage dimensions work in a child with a parent with viewport unit dimensions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS window.outerWidth is 800
PASS window.outerHeight is 600
PASS getComputedStyle(container).width is '400px'
PASS getComputedStyle(container).height is '300px'
PASS getComputedStyle(percent).width is '200px'
PASS getComputedStyle(percent).height is '150px'
PASS window.outerWidth is 900
PASS window.outerHeight is 600
PASS getComputedStyle(container).width is '400px'
PASS getComputedStyle(container).height is '300px'
PASS getComputedStyle(percent).width is '200px'
PASS getComputedStyle(percent).height is '150px'
PASS window.outerWidth is 900
PASS window.outerHeight is 640
PASS getComputedStyle(container).width is '400px'
PASS getComputedStyle(container).height is '300px'
PASS getComputedStyle(percent).width is '200px'
PASS getComputedStyle(percent).height is '150px'
PASS window.outerWidth is 500
PASS window.outerHeight is 640
PASS getComputedStyle(container).width is '400px'
PASS getComputedStyle(container).height is '300px'
PASS getComputedStyle(percent).width is '200px'
PASS getComputedStyle(percent).height is '150px'
PASS window.outerWidth is 800
PASS window.outerHeight is 600
PASS getComputedStyle(container).width is '400px'
PASS getComputedStyle(container).height is '300px'
PASS getComputedStyle(percent).width is '200px'
PASS getComputedStyle(percent).height is '150px'
